Update doubles to decimals to avoid precision errors

+++ DGavian.GapFiller.UnitTests/TestHelpers/TestData.cs
@@ -9,6 +9,6 @@ namespace DGavian.GapFiller.UnitTests.TestHelpers
 {
     public class TestData : IOffset
     {
-        public double Offset { get; set; }
+        public decimal Offset { get; set; }
     }
 }
